+++ modules/cf-core/src/methods/app-instance/get-app-instance.ts
@@ -2,7 +2,7 @@ import { MethodNames, MethodParams, MethodResults } from "@connext/types";
 
 import { NO_APP_IDENTITY_HASH_TO_GET_DETAILS, NO_APP_INSTANCE_FOR_GIVEN_HASH } from "../../errors";
 import { RequestHandler } from "../../request-handler";
-import { jsonRpcMethod } from "../../rpc-server";
+import { jsonRpcMethod } from "../../rpc-router";
 
 import { NodeController } from "../controller";
