+++ modules/cf-core/src/methods/app-instance/get-token-indexed-free-balances.ts
@@ -3,7 +3,7 @@ import { MethodNames, MethodParams, MethodResults } from "@connext/types";
 import { NO_STATE_CHANNEL_FOR_MULTISIG_ADDR } from "../../errors";
 import { StateChannel } from "../../models";
 import { RequestHandler } from "../../request-handler";
-import { jsonRpcMethod } from "../../rpc-server";
+import { jsonRpcMethod } from "../../rpc-router";
 
 import { NodeController } from "../controller";
