+++ modules/cf-core/src/methods/app-instance/install.ts
@@ -17,8 +17,7 @@ import {
 import { ProtocolRunner } from "../../machine";
 import { StateChannel } from "../../models";
 import { RequestHandler } from "../../request-handler";
-import { RpcRouter } from "../../rpc-router";
-import { jsonRpcMethod } from "../../rpc-server";
+import { jsonRpcMethod, RpcRouter } from "../../rpc-router";
 
 import { NodeController } from "../controller";
