+++ modules/cf-core/src/methods/controller.ts
@@ -1,7 +1,7 @@
 import { MethodName, MethodParam, MethodResult } from "@connext/types";
 import { logTime } from "@connext/utils";
 
-import { Controller } from "../rpc-server";
+import { Controller } from "../rpc-router";
 import { RequestHandler } from "../request-handler";
 import { StateChannel } from "../models/state-channel";
