+++ modules/cf-core/src/methods/state-channel/get.ts
@@ -2,7 +2,7 @@ import { MethodNames, MethodParams, MethodResults } from "@connext/types";
 
 import { NO_STATE_CHANNEL_FOR_MULTISIG_ADDR } from "../../errors";
 import { RequestHandler } from "../../request-handler";
-import { jsonRpcMethod } from "../../rpc-server";
+import { jsonRpcMethod } from "../../rpc-router";
 
 import { NodeController } from "../controller";
